[JT-63] refactor: Episode 도메인 서비스 분리
